+++ sdk/keyvault/azsecrets/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Breaking Changes
 * Removed `RawResponse` from pollers
 * Polling operations return a Poller struct directly instead of a Response envelope
+* Removed `ToPtr` methods
 
 ## 0.6.0 (2022-03-08)